pyproject.toml
[tool.poetry]
name = "bearparse"
version = "0.3.0"
description = "A custom argument parser for non-standard arguments. Useful for Appworx"
authors = ["Zevaryx <zevaryx@gmail.com>"]
license = "GPL-3.0-or-later"

[tool.poetry.dependencies]
python = "^3.8"
attrs = "^21.2.0"
PyYAML = "^5.4.1"

[tool.poetry.dev-dependencies]
pytest = "^6.2.5"
pytest-cov = "^2.12.1"

[build-system]
requires = ["poetry-core>=1.0.0"]
build-backend = "poetry.core.masonry.api"
